refactor(web-js-sdk): move hasLoginIndicator to withLoggedInIndicator
The function only reads keys owned by withLoggedInIndicator (DSLI) and
withLastLoggedInUser (lastUser bootstrap), so it belongs alongside the
enhancer that writes those keys rather than under withPersistTokens.
