+++ packages/sdks/web-js-sdk/test/persistTokens.test.ts
@@ -3,7 +3,6 @@ import {
   getIdToken,
   getSessionToken,
   getRefreshToken,
-  hasLoginIndicator,
 } from '../src/enhancers/withPersistTokens/helpers';
 import createSdk from '../src/index';
 import { authInfo, oidcAuthInfo } from './mocks';
@@ -1213,33 +1212,6 @@ describe('persistTokens', () => {
       localStorage.setItem('DSI', 'id-token-1');
 
       expect(getIdToken()).toEqual('id-token-1');
-    });
-  });
-
-  describe('hasLoginIndicator', () => {
-    beforeEach(() => {
-      localStorage.removeItem('DSLI');
-      localStorage.removeItem('dls_last_user_login_id');
-    });
-
-    it('returns true when DSLI is set', () => {
-      localStorage.setItem('DSLI', '1');
-      expect(hasLoginIndicator()).toBe(true);
-    });
-
-    it('returns true when only the lastUser bootstrap key is set', () => {
-      localStorage.setItem('dls_last_user_login_id', 'someone@example.com');
-      expect(hasLoginIndicator()).toBe(true);
-    });
-
-    it('returns true when both DSLI and lastUser are set', () => {
-      localStorage.setItem('DSLI', '1');
-      localStorage.setItem('dls_last_user_login_id', 'someone@example.com');
-      expect(hasLoginIndicator()).toBe(true);
-    });
-
-    it('returns false when neither key is set', () => {
-      expect(hasLoginIndicator()).toBe(false);
     });
   });
